Updated failing execution tests containing push.

+++ test/libsolidity/semanticTests/viaYul/array_storage_index_boundary_test.sol
@@ -1,6 +1,6 @@
 contract C {
     uint[] storageArray;
-    function test_boundary_check(uint256 len, uint256 access) public returns (uint256)
+    function test_boundary_check(uint len, uint access) public returns (uint)
     {
         while(storageArray.length < len)
             storageArray.push();
@@ -12,12 +12,12 @@ contract C {
 // ====
 // compileViaYul: also
 // ----
-// test_boundary_check(uint256,uint256): 10, 11 -> FAILURE, hex"4e487b71", 0x32
-// test_boundary_check(uint256,uint256): 10, 9 -> 0
-// test_boundary_check(uint256,uint256): 1, 9 -> FAILURE, hex"4e487b71", 0x32
-// test_boundary_check(uint256,uint256): 1, 1 -> FAILURE, hex"4e487b71", 0x32
-// test_boundary_check(uint256,uint256): 10, 10 -> FAILURE, hex"4e487b71", 0x32
-// test_boundary_check(uint256,uint256): 256, 256 -> FAILURE, hex"4e487b71", 0x32
-// test_boundary_check(uint256,uint256): 256, 255 -> 0
-// test_boundary_check(uint256,uint256): 256, 0xFFFF -> FAILURE, hex"4e487b71", 0x32
-// test_boundary_check(uint256,uint256): 256, 2 -> 0
+// test_boundary_check(uint,uint): 10, 11 -> FAILURE, 255
+// test_boundary_check(uint,uint): 10, 9 -> 0
+// test_boundary_check(uint,uint): 1, 9 -> FAILURE, 255
+// test_boundary_check(uint,uint): 1, 1 -> FAILURE, 255
+// test_boundary_check(uint,uint): 10, 10 -> FAILURE, 255
+// test_boundary_check(uint,uint): 256, 256 -> FAILURE, 255
+// test_boundary_check(uint,uint): 256, 255 -> 0
+// test_boundary_check(uint,uint): 256, 0xFFFF -> FAILURE, 255
+// test_boundary_check(uint,uint): 256, 2 -> 0
